+++ src/providers/index.ts
@@ -1,2 +1,3 @@
 export * from "./consumerVirtualTextDocumentProvider";
+export * from "./consumerTableViewProvider";
 export * from "./outputChannelProvider";